Reduced text size to 1/5, this lines up better with the original font and normal text sizes are achieved with scales closer to 1. PrepareDrawSurface no longer sets background color. Changed height and with members of Texture to float rather than int to accomodate the smaller texture sizes

+++ DrawLib/Texture.cs
@@ -72,14 +72,14 @@ namespace IngameScript
 			/// <summary>
 			/// The actual width of the Texture in pixels (including scaling)
 			/// </summary>
-			public int Width { get { return (int)(width * Scale.X); } set { Scale.X = width / Width; } }
-			private int width;
+			public float Width { get { return width * Scale.X; } set { Scale.X = width / Width; } }
+			private float width;
 
 			/// <summary>
 			/// The actual height of the Texture in pixels (including scaling)
 			/// </summary>
-			public int Height { get { return (int)(height * Scale.Y); } set { Scale.Y = height / Height; } }
-			private int height;
+			public float Height { get { return height * Scale.Y; } set { Scale.Y = height / Height; } }
+			private float height;
 
 			/// <summary>
 			/// The colour that gets assigned to all sprites classed as "background"
@@ -107,7 +107,7 @@ namespace IngameScript
 			/// <param name="_name">Sets the display name. If not specified, "texture" will be used as default</param>
 			/// <param name="_sprites">A list of sprites, if nothing or null is passed, a new empty List is initialized instead</param>
 			/// <param name="_rotation">A default Value for RotationOrScale. Keep in mind that Text sprites can not be Rotated, and will be scaled instead</param>
-			public Texture(string _name = "texture", List<MyTuple<MySprite,Texture.ColorSlot>> _sprites = null, float _rotation = 0f, int _width = 0, int _height = 0, Dictionary<string,Color> _colors = null)
+			public Texture(string _name = "texture", List<MyTuple<MySprite,Texture.ColorSlot>> _sprites = null, float _rotation = 0f, float _width = 0, float _height = 0, Dictionary<string,Color> _colors = null)
 			{
 				if (_sprites == null)
 					sprites = new List<MyTuple<MySprite, Texture.ColorSlot>>();
